Move all handlers to handlers.py
- Rename strategies.py to handlers.py and move all handlers to it:
