+++ clarity_ext/service/dilution/service.py
@@ -525,136 +525,6 @@ class RobotSettings(object):
             file_ext=self.file_ext)
 
 
-class TransferBatchHandlerBase(object):
-    __metaclass__ = abc.ABCMeta
-
-    def __init__(self, validation_service):
-        self.validation_service = validation_service
-
-    @abc.abstractmethod
-    def needs_split(self, transfer, dilution_settings, robot_settings):
-        pass
-
-    def execute(self, transfer_batch, dilution_settings, robot_settings, strategy):
-        """
-        Returns one or two transfer_batches, based on rules. Can be used to split a transfer_batch into original and
-        temporary transfer_batches
-        """
-        split = [t for t in transfer_batch.transfers if self.needs_split(t, dilution_settings, robot_settings)]
-        no_split = [t for t in transfer_batch.transfers if t not in split]
-
-        for transfer in split:
-            # This may look strange, but for now we want to handle this like a validation exception
-            val = NeedsBatchSplit(transfer)
-            self.validation_service.handle_single_validation(val)
-
-        if len(split) > 0:
-            return self.split_transfer_batch(split, no_split, strategy, robot_settings)
-        else:
-            # No split was required
-            return TransferBatchCollection(transfer_batch)
-
-    def split_transfer_batch(self, split, no_split, strategy, robot_settings):
-        first_transfers = list(self.calculate_split_transfers(split))
-        temp_transfer_batch = TransferBatch(first_transfers, robot_settings, depth=1, is_temporary=True)
-        strategy.calculate_transfer_volumes(temp_transfer_batch)
-        second_transfers = list()
-
-        # We need to create a new transfers list with:
-        #  - target_location should be the original target location
-        #  - source_location should also bet the original source location
-        for temp_transfer in temp_transfer_batch.transfers:
-            # NOTE: It's correct that the source_location is being set to the target_location here:
-            new_transfer = SingleTransfer(temp_transfer.target_conc, temp_transfer.target_vol,
-                                          temp_transfer.original.target_conc, temp_transfer.original.target_vol,
-                                          source_location=temp_transfer.target_location,
-                                          target_location=temp_transfer.original.target_location)
-
-            # In the case of a split TransferBatch, only the secondary transfer should update source volume:
-            new_transfer.should_update_source_vol = False
-            second_transfers.append(new_transfer)
-
-        # Add other transfers just as they were:
-        second_transfers.extend(no_split)
-
-        final_transfer_batch = TransferBatch(second_transfers, robot_settings, depth=1)
-        strategy.calculate_transfer_volumes(final_transfer_batch)
-
-        # For the analytes requiring splits
-        return TransferBatchCollection(temp_transfer_batch, final_transfer_batch)
-
-    def calculate_split_transfers(self, original_transfers):
-        # For each target well, we need to push this to a temporary plate:
-
-        # First we need a map from the actual target plates to temp plates:
-        map_target_container_to_temp = dict()
-        for transfer in original_transfers:
-            target_container = transfer.target_location.container
-            if target_container not in map_target_container_to_temp:
-                temp_container = Container.create_from_container(target_container)
-                temp_container.id = "temp{}".format(len(map_target_container_to_temp) + 1)
-                temp_container.name = temp_container.id
-                map_target_container_to_temp[target_container] = temp_container
-
-        for transfer in original_transfers:
-            temp_target_container = map_target_container_to_temp[transfer.target_location.container]
-            # TODO: Copy the source location rather than using the original?
-
-            # Create a temporary analyte representing the new one on the temp plate:
-            temp_analyte = copy.copy(transfer.source_location.artifact)
-            temp_analyte.id += "-temp"
-            temp_analyte.name += "-temp"
-            temp_target_location = temp_target_container.set_well(
-                transfer.target_location.position,
-                temp_analyte)
-
-            # TODO: This should be defined by a rule provided by the inheriting class
-            static_sample_volume = 4
-            static_buffer_volume = 36
-            transfer_copy = SingleTransfer(transfer.source_conc,
-                                           transfer.source_vol,
-                                           transfer.source_conc / 10.0,
-                                           static_buffer_volume + static_sample_volume,
-                                           transfer.source_location,
-                                           temp_target_location)
-            transfer_copy.is_primary = False
-
-            # In this case, we'll hardcode the values according to the lab's specs:
-            transfer_copy.pipette_sample_volume = static_sample_volume
-            transfer_copy.pipette_buffer_volume = static_buffer_volume
-            transfer_copy.original = transfer
-
-            yield transfer_copy
-
-
-class TransferHandlerBase(object):
-    __metaclass__ = abc.ABCMeta
-
-    def __init__(self, validation_service):
-        self.validation_service = validation_service
-
-    @abc.abstractmethod
-    def needs_row_split(self, transfer, dilution_settings, robot_settings):
-        pass
-
-    @abc.abstractmethod
-    def split_single_transfer(self, transfer, robot_settings):
-        pass
-
-    def execute(self, transfer_batch, dilution_settings, robot_settings):
-        require_row_split = [t for t in transfer_batch.transfers
-                             if self.needs_row_split(t, dilution_settings, robot_settings)]
-        for transfer in require_row_split:
-            # TODO: Looks weird to push the action to the validation handler like this, but fits right now.
-            val = NeedsRowSplit(transfer)
-            self.validation_service.handle_single_validation(val)
-
-        for transfer in require_row_split:
-            split_transfers = self.split_single_transfer(transfer, robot_settings)
-            transfer_batch.transfers.remove(transfer)
-            transfer_batch.transfers.extend(split_transfers)
-
-
 class DilutionValidatorBase(object):
     """
     Validates transfer objects that are to be diluted. Inherit from this object to support behavior
